requirements.txt
cohere
guardrails-ai
pydantic
